requirements.txt
numpy 
opencv-python 
torch 
torchvision 
scipy 
open3d 
ultralytics 
onnxruntime-gpu
tqdm
matplotlib